pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "discord-py>=2.6.3",
    "python-dotenv>=1.1.1",
]
